requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
openpyxl>=3.1.0
pandas>=2.0.0
PyYAML>=6.0
pymysql>=1.1.0
